+++ Package.swift
@@ -12,6 +12,11 @@ let package = Package(
         .library(name: "Sora", targets: ["Sora"]),
         .library(name: "WebRTC", targets: ["WebRTC"]),
     ],
+    dependencies: [
+        // 開発用依存関係
+        // SwfitLint 公式で推奨されている SwfitLintPlugins を利用する
+        .package(url: "https://github.com/SimplyDanny/SwiftLintPlugins", from: "0.58.2")
+    ],
     targets: [
         .binaryTarget(
             name: "WebRTC",